% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{cardinal_sine_model}
\alias{cardinal_sine_model}
\title{Cardinal Sine Model}
\usage{
cardinal_sine_model(h, sill)
}
\arguments{
\item{h}{Distance.}

\item{sill}{Sill value.}
}
\value{
Computed value based on the cardinal sine model.
}
\description{
Calculate the cardinal sine model based on the given parameters.
}
