% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{cauchy_model}
\alias{cauchy_model}
\title{Cauchy Model}
\usage{
cauchy_model(h, nugget, sill, range, a)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{range}{Range value.}

\item{a}{Additional parameter.}
}
\value{
Computed value based on the Cauchy model.
}
\description{
Calculate the Cauchy model based on the given parameters.
}
