% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mtrx.R
\name{cor_mtrx}
\alias{cor_mtrx}
\title{correlation matrix}
\usage{
cor_mtrx(data)
}
\arguments{
\item{data}{A dataset of class data.frame. The data should contain 'C_X' and 'C_Y' columns
representing the x and y coordinates of the data points and excludes any rank,
ID, or column not for analysis.}
}
\value{
a correlation matrix as matrix object
}
\description{
correlation matrix
}
\examples{
data("Klovan_Row80")
cor_mtrx(range_transform(Klovan_Row80)) # view correlation matrix
corMtrx <- cor_mtrx(Klovan_Row80) # save correlation matrix as object
}
