% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{correlation_plot}
\alias{correlation_plot}
\title{Correlation Plot}
\usage{
correlation_plot(data)
}
\arguments{
\item{data}{A dataset of class data.frame. The data should contain 'C_X' and 'C_Y' columns
representing the x and y coordinates of the data points and excludes any rank,
ID, or column not for analysis, see README for details}
}
\value{
a correlation plot displaying correlation coefficients, the scatter plots with a regression line and, the variable histograms in a 10 x 10 table.
}
\description{
The correlation plot is a summary showing the relationship among variables.The plot below is a 10 x 10 table where each variable is plotted against every other variable.In the top half of the table, the correlation coefficients are displayed.  In the bottom half, the scatter plots are shown along with a regression line.  Down the diagonal axis, the variable histograms are show.
}
\examples{
\donttest{
data("Klovan_Row80")
correlation_plot(Klovan_Row80)
}
}
