% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{cosinus_model}
\alias{cosinus_model}
\title{Cosinus Model}
\usage{
cosinus_model(h, sill, a)
}
\arguments{
\item{h}{Distance.}

\item{sill}{Sill value.}

\item{a}{Additional parameter.}
}
\value{
Computed value based on the cosinus model.
}
\description{
Calculate the cosinus model based on the given parameters.
}
