% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{exponential_model}
\alias{exponential_model}
\title{Exponential Model}
\usage{
exponential_model(h, nugget, sill, range)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{range}{Range value.}
}
\value{
Computed value based on the exponential model.
}
\description{
Calculate the exponential model based on the given parameters.
}
