% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{gamma_model}
\alias{gamma_model}
\title{Gamma Model}
\usage{
gamma_model(h, nugget, sill, range, a)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{range}{Range value.}

\item{a}{Additional parameter.}
}
\value{
Computed value based on the gamma model.
}
\description{
Calculate the gamma model based on the given parameters.
}
