% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{gaussian_model}
\alias{gaussian_model}
\title{Gaussian Model}
\usage{
gaussian_model(h, nugget, sill, range)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{range}{Range value.}
}
\value{
Computed value based on the Gaussian model.
}
\description{
Calculate the Gaussian model based on the given parameters.
}
