% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klovan_Row80.R
\docType{data}
\name{Klovan_Row80}
\alias{Klovan_Row80}
\title{Klovan mining dataset}
\format{
A data frame with observations of various geological variables
@item rank: Rank of the observation
@item C_X: X-coordinate of the observation
@item C_Y: Y-coordinate of the observation
@item P_Elong: Elongation of the geological feature
@item P_Fe: Iron content in the geological feature
@item P_Fold: Folding factor of the geological feature
@item P_Fract: Fracturing level of the geological feature
@item P_Mg: Magnesium content in the geological feature
@item P_Na: Sodium content in the geological feature
@item P_Space: Spatial arrangement factor of the geological feature
@item P_Sulfide: Sulfide content in the geological feature
@item P_Veins: Veining factor of the geological feature
@item P_XLSize: Size of the geological feature
}
\usage{
Klovan_Row80
}
\description{
The \code{klovan} dataset provides exploration data for geostatistics and analytical methods. The region has undergone extensive geological investigation, revealing the position of a known ore body.
}
\keyword{FA}
\keyword{PCA}
\keyword{datascience}
\keyword{dataset}
\keyword{geology}
\keyword{geostatistics}
