% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_outliers.R
\name{manage_outliers}
\alias{manage_outliers}
\title{Outlier Management}
\usage{
manage_outliers(data, property)
}
\arguments{
\item{data}{A dataset of class data.frame.}

\item{property}{A string representing the property on which the range transformation is based.}
}
\value{
The input dataset, supplemented with a new Boolean column. TRUE signifies a high likelihood of an outlier, while FALSE signifies a low likelihood.
}
\description{
This function appends a new column to the input data, marking potential outliers.
Once identified, these outliers can either be removed or imputed.
}
\examples{
data("Klovan_2D_all_outlier")
manage_outliers(Klovan_2D_all_outlier, "P_Mg")

}
