% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{matern_model}
\alias{matern_model}
\title{Matern Model}
\usage{
matern_model(h, nugget, sill, range, a)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{range}{Range value.}

\item{a}{Shape parameter.}
}
\value{
Computed value based on the Matern model.
}
\description{
Calculate the Matern model based on the given parameters.
}
