% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{order_1_gc_model}
\alias{order_1_gc_model}
\title{Order-1 G.C. (General Covariance) Model}
\usage{
order_1_gc_model(h, nugget, sill, range)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{range}{Range value.}
}
\value{
Computed value based on the order-1 G.C. model.
}
\description{
Calculate the order-1 G.C. model based on the given parameters.
}
