% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{power_model}
\alias{power_model}
\title{Power Model}
\usage{
power_model(h, nugget, sill, a)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{a}{Power parameter.}
}
\value{
Computed value based on the power model.
}
\description{
Calculate the power model based on the given parameters.
}
