% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{print_model_names}
\alias{print_model_names}
\title{Print Model Names}
\usage{
print_model_names()
}
\value{
NULL (This function is used for printing the model names only.)
}
\description{
This function prints the names of the predefined model functions.
}
\examples{
print_model_names()

}
