% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{quadratic_exponential_model}
\alias{quadratic_exponential_model}
\title{Quadratic Exponential Model}
\usage{
quadratic_exponential_model(h, nugget, sill, range, a)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{range}{Range value.}

\item{a}{Additional parameter.}
}
\value{
Computed value based on the quadratic exponential model.
}
\description{
Calculate the quadratic exponential model based on the given parameters.
}
