% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_transform.R
\name{range_transform}
\alias{range_transform}
\title{range transform}
\usage{
range_transform(data)
}
\arguments{
\item{data}{A dataset of class data.frame. The data should contain 'C_X' and 'C_Y' columns
representing the x and y coordinates of the data points and excludes any rank,
ID, or column not for analysis, see README for details}
}
\value{
a range transformed version of a klovan dataset.
}
\description{
Normalize the data using a 'Range' transform .
In the returned data table, note that in each column of the normalized Data Table, the variables will range from 0 to 1.
}
\examples{
data("Klovan_Row80")
T_Klovan <- range_transform(Klovan_Row80)

}
