% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{spherical_model}
\alias{spherical_model}
\title{Spherical Model}
\usage{
spherical_model(h, nugget, sill, range)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{range}{Range value.}
}
\value{
Computed value based on the spherical model.
}
\description{
Calculate the spherical model based on the given parameters.
}
