% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{stable_model}
\alias{stable_model}
\title{Stable Model}
\usage{
stable_model(h, nugget, sill, range, a)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{range}{Range value.}

\item{a}{Additional parameter.}
}
\value{
Computed value based on the stable model.
}
\description{
Calculate the stable model based on the given parameters.
}
