% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogramModels.R
\name{triangle_model}
\alias{triangle_model}
\title{Triangle Model}
\usage{
triangle_model(h, nugget, sill, range)
}
\arguments{
\item{h}{Distance.}

\item{nugget}{Nugget effect.}

\item{sill}{Sill value.}

\item{range}{Range value.}
}
\value{
Computed value based on the triangle model.
}
\description{
Calculate the triangle model based on the given parameters.
}
